+++ lib/features/family/features/reflect/bloc/interview_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:givt_app/features/family/features/reflect/domain/models/game_profile.dart';
 import 'package:givt_app/features/family/features/reflect/domain/reflect_and_share_repository.dart';
 import 'package:givt_app/features/family/features/reflect/presentation/models/interview_custom.dart';
-import 'package:givt_app/features/family/features/reflect/presentation/models/record_answer_uimodel.dart';
+import 'package:givt_app/features/family/features/reflect/presentation/models/interview_uimodel.dart';
 import 'package:givt_app/shared/bloc/base_state.dart';
 import 'package:givt_app/shared/bloc/common_cubit.dart';
